[cmd] move all flags to flags.go - one file
